requirements.txt
Flask
Flask-cors
pyinstaller
esptool
pyserial